fixup for nosetest-ification
- Add --with-xunit to debian package run
- Fixups:
  - Move utils.py to __init__.py
  - Fix nosetests target in src/Makefile
