+++ nosetests/test_030_comp_or2.py
@@ -1,6 +1,4 @@
-#!/usr/bin/env python
-
-from utils import RTAPITestCase
+from . import RTAPITestCase
 from nose.tools import assert_equal, assert_true, assert_false
 
 from machinekit import hal
